#143 Remove VerificationMail's Interface
